Rename hydra_cfg to hydraconf
